+++ src/parser/model/organizational-unit-resource.ts
@@ -34,6 +34,8 @@ export class OrganizationalUnitResource extends Resource {
 
         this.organizationalUnitName = this.props.OrganizationalUnitName;
 
+        super.throwForNonRef(this.props.OrganizationalUnits, 'OrganizationalUnits');
+        super.throwForNonRef(this.props.Accounts, 'Accounts');
         super.throwForUnknownAttributes(resource, id, 'Type', 'Properties');
         super.throwForUnknownAttributes(this.props, id, 'OrganizationalUnitName', 'Accounts', 'ServiceControlPolicies', 'OrganizationalUnits');
     }